refactor(cli): convert cli.rs to cli/ module and extract runner.rs
- Rename src/cli.rs to src/cli/mod.rs (content unchanged)
- Add module declarations for config_resolver and runner in mod.rs
- Create src/cli/config_resolver.rs as empty stub
- Extract display_banner, resolve_suggest_fix, validate_project_path
  from main.rs into src/cli/runner.rs
- Move corresponding tests to runner.rs
- Update main.rs to import functions from cli::runner

Part of #315. Closes #353
