+++ src/cli/mod.rs
@@ -1,3 +1,6 @@
+pub mod config_resolver;
+pub mod runner;
+
 use clap::Parser;
 
 use crate::application::dto::OutputFormat;
